requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
transliterate==1.10.2
tzdata==2025.2
